rename dashboard -> cloud
