+++ packages/server/test/integration/cypress_spec.js
@@ -23,9 +23,9 @@ const ciProvider = require(`../../lib/util/ci_provider`)
 const settings = require(`../../lib/util/settings`)
 const Windows = require(`../../lib/gui/windows`)
 const interactiveMode = require(`../../lib/modes/interactive`)
-const api = require(`../../lib/dashboard/api`)
+const api = require(`../../lib/cloud/api`)
 const cwd = require(`../../lib/cwd`)
-const user = require(`../../lib/dashboard/user`)
+const user = require(`../../lib/cloud/user`)
 const cache = require(`../../lib/cache`)
 const errors = require(`../../lib/errors`)
 const cypress = require(`../../lib/cypress`)
