+++ packages/server/test/unit/dashboard/api_spec.js
@@ -6,10 +6,10 @@ const {
   agent,
 } = require('@packages/network')
 const pkg = require('@packages/root')
-const api = require('../../../lib/dashboard/api')
+const api = require('../../../lib/cloud/api')
 const cache = require('../../../lib/cache')
 const errors = require('../../../lib/errors')
-const machineId = require('../../../lib/dashboard/machine_id')
+const machineId = require('../../../lib/cloud/machine_id')
 const Promise = require('bluebird')
 
 const API_BASEURL = 'http://localhost:1234'
@@ -23,7 +23,7 @@ const makeError = (details = {}) => {
   return _.extend(new Error(details.message || 'Some error'), details)
 }
 
-describe('lib/dashboard/api', () => {
+describe('lib/cloud/api', () => {
   beforeEach(() => {
     nock(API_BASEURL)
     .matchHeader('x-route-version', '2')
