+++ packages/server/test/unit/errors_spec.js
@@ -1,6 +1,6 @@
 require('../spec_helper')
 
-const exception = require(`../../lib/dashboard/exception`)
+const exception = require(`../../lib/cloud/exception`)
 const chalk = require('chalk')
 const errors = require('../../lib/errors')
